requirements.txt
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
boto==2.48.0
celery==4.3.0
configparser==3.5.0
dj-database-url==0.5.0
Django==2.0.7
django-crispy-forms==1.7.2
django-pagedown==1.0.4
django-phonenumber-field==3.0.1
django-registration-redux==2.4
entrypoints==0.3
flake8==3.7.8
gunicorn==19.8.1
importlib-metadata==0.23
kombu==4.6.5
mccabe==0.6.1
more-itertools==7.2.0
phonenumbers==8.10.21
Pillow==6.0.0
psycopg2-binary==2.7.5
pycodestyle==2.5.0
pyflakes==2.1.1
python-stdnum==1.11
pytils==0.3
pytz==2019.3
redis==3.3.11
vine==1.3.0
zipp==0.6.0
